(PC-29655)[PRO] refactor: remove create/update individual offer adapters
