+++ pro/src/screens/IndividualOffer/InformationsScreen/serializePatchOffer.ts
