+++ pro/src/screens/IndividualOffer/InformationsScreen/serializePostOffer.ts
