+++ pro/src/screens/IndividualOffer/PriceCategoriesScreen/form/submitToApi.ts
@@ -1,6 +1,5 @@
 import { api } from 'apiClient/api'
 import { GetIndividualOfferResponseModel } from 'apiClient/v1'
-import { updateIndividualOffer } from 'core/Offers/adapters/updateIndividualOffer/updateIndividualOffer'
 
 import { computeInitialValues } from './computeInitialValues'
 import { serializePriceCategories } from './serializePriceCategories'
@@ -11,13 +10,12 @@ export const submitToApi = async (
   offer: GetIndividualOfferResponseModel,
   resetForm: PriceCategoryFormik['resetForm']
 ) => {
-  const { isOk: isOfferOk, message: offerMessage } =
-    await updateIndividualOffer({
-      offerId: offer.id,
-      serializedOffer: { isDuo: values.isDuo },
-    })
-  if (!isOfferOk) {
-    throw new Error(offerMessage)
+  try {
+    await api.patchOffer(offer.id, { isDuo: values.isDuo })
+  } catch {
+    throw new Error(
+      'Une erreur est survenue lors de la création de votre offre'
+    )
   }
 
   try {
